% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rm_0s_by_col}
\alias{rm_0s_by_col}
\title{To remove columns that only have zero values}
\usage{
rm_0s_by_col(dado)
}
\arguments{
\item{dado}{Is the input data.frame or matrix in which the function is going to remove columns that only have zeroes}
}
\description{
Description of the function
}
